+++ apps/examples/CHANGELOG.md
@@ -1,5 +1,14 @@
 # examples
 
+## 0.1.74
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/react@0.1.8
+  - @orderly.network/hooks@1.0.5
+  - @orderly.network/core@0.1.4
+
 ## 0.1.73
 
 ### Patch Changes