+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @orderly.network/hooks
 
+## 1.0.5
+
+### Patch Changes
+
+- fix style bugs;
+- Updated dependencies
+  - @orderly.network/types@0.1.4
+  - @orderly.network/core@0.1.4
+  - @orderly.network/perp@1.0.4
+  - @orderly.network/net@1.0.91
+
 ## 1.0.4
 
 ### Patch Changes