+++ packages/onboard/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @orderly.network/onboard
 
+## 0.0.5
+
+### Patch Changes
+
+- fix style bugs;
+- Updated dependencies
+  - @orderly.network/hooks@1.0.5
+  - @orderly.network/types@0.1.4
+
 ## 0.0.4
 
 ### Patch Changes