+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @orderly.network/types
 
+## 0.1.4
+
+### Patch Changes
+
+- fix style bugs;
+
 ## 0.1.3
 
 ### Patch Changes